requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-nvidia-ai-endpoints>=0.3.0
faiss-cpu
pandas
python-dotenv
pydantic